requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
flake8==7.1.1
idna==3.10
mccabe==0.7.0
pycodestyle==2.12.1
pyflakes==3.2.0
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
